+++ module/coc7.js
@@ -1,23 +1,21 @@
-/* global $, CONFIG, foundry, game, Hooks, ItemDirectory */
+/* global Hooks */
 import '../styles/system/index.less'
-import CoC7ActiveEffect from './active-effect.js'
 import { CoC7NPCSheet } from './actors/sheets/npc-sheet.js'
 import { CoC7CreatureSheet } from './actors/sheets/creature-sheet.js'
 import { CoC7CharacterSheet } from './actors/sheets/character.js'
 import { CoC7Chat } from './chat.js'
 import { CoC7Combat } from './combat.js'
-import { COC7 } from './config.js'
-import { Updater } from './updater.js'
 import { CoC7Utilities } from './utilities.js'
-import { CoC7Check } from './check.js'
 import { CoC7Menu } from './menu.js'
 import { CoC7Canvas } from './apps/canvas.js'
 import { CoC7Hooks } from './hooks/index.js'
 import * as DiceBot from './dicebot.js'
 import { CoC7ChaseSheet } from './items/chase/sheet.js'
 import { CoC7Socket } from './hooks/socket.js'
-import { CoC7SystemSocket } from './apps/coc7-system-socket.js'
 import { DropActorSheetData } from './hooks/drop-actor-sheet-data.js'
+import ChangeSidebarTab from './hooks/change-sidebar-tab.js'
+import CreateActiveEffect from './hooks/create-active-effect.js'
+import DeleteActiveEffect from './hooks/delete-active-effect.js'
 import DrawNote from './hooks/draw-note.js'
 import GetJournalSheetHeaderButtons from './hooks/get-journal-sheet-header-buttons.js'
 import GetMacroConfigHeaderButtons from './hooks/get-macro-config-header-buttons.js'
@@ -26,17 +24,18 @@ import GetRollTableConfigHeaderButtons from './hooks/get-roll-table-config-heade
 import GetSceneConfigHeaderButtons from './hooks/get-scene-config-header-buttons.js'
 import Init from './hooks/init.js'
 import PopOutLoaded from './hooks/popout-loaded.js'
+import Ready from './hooks/ready.js'
 import RenderActorDirectory from './hooks/render-actor-directory.js'
 import RenderCoC7JournalSheet from './hooks/render-coc7-journal-sheet.js'
 import RenderCompendiumDirectory from './hooks/render-compendium-directory.js'
 import RenderJournalTextPageSheet from './hooks/render-journal-text-page-sheet.js'
 import RenderSettings from './hooks/render-settings.js'
 import RenderSettingsConfig from './hooks/render-settings-config.js'
+import Setup from './hooks/setup.js'
 
 // Card init
 import { initECC } from './common/chatcardlib/src/chatcardlib.js'
 import { ChaseObstacleCard } from './chat/cards/chase-obstacle.js'
-import { CoC7ContextMenu } from './context-menu.js'
 
 Hooks.once('init', Init)
 
@@ -51,65 +50,7 @@ CoC7Hooks.listen()
 
 Hooks.once('socketlib.ready', CoC7Socket)
 
-Hooks.once('setup', function () {
-  // Localize CONFIG objects once up-front
-  const toLocalize = [
-    'spellProperties',
-    'bookType',
-    'talentType',
-    'occupationProperties',
-    'statusType'
-  ]
-
-  for (const o of toLocalize) {
-    const localized = Object.entries(COC7[o]).map(e => {
-      return [e[0], game.i18n.localize(e[1])]
-    })
-    COC7[o] = localized.reduce((obj, e) => {
-      obj[e[0]] = e[1]
-      return obj
-    }, {})
-  }
-
-  let effectIndex = CONFIG.statusEffects.findIndex(
-    t => t.id === COC7.status.dead
-  )
-  if (effectIndex !== -1) {
-    CONFIG.statusEffects[effectIndex].icon =
-      'systems/CoC7/assets/icons/tombstone.svg'
-  }
-  effectIndex = CONFIG.statusEffects.findIndex(
-    t => t.id === COC7.status.unconscious
-  )
-  if (effectIndex !== -1) {
-    CONFIG.statusEffects[effectIndex].icon =
-      'systems/CoC7/assets/icons/knocked-out-stars.svg'
-  }
-  // FoundryVTT v10
-  const effectNameKey = (!foundry.utils.isNewerVersion(game.version, '11') ? 'label' : 'name')
-  CONFIG.statusEffects.unshift(
-    {
-      id: COC7.status.tempoInsane,
-      [effectNameKey]: 'CoC7.BoutOfMadnessName',
-      icon: 'systems/CoC7/assets/icons/hanging-spider.svg'
-    },
-    {
-      id: COC7.status.indefInsane,
-      [effectNameKey]: 'CoC7.InsanityName',
-      icon: 'systems/CoC7/assets/icons/tentacles-skull.svg'
-    },
-    {
-      id: COC7.status.criticalWounds,
-      [effectNameKey]: 'CoC7.CriticalWounds',
-      icon: 'systems/CoC7/assets/icons/arm-sling.svg'
-    },
-    {
-      id: COC7.status.dying,
-      [effectNameKey]: 'CoC7.Dying',
-      icon: 'systems/CoC7/assets/icons/heart-beats.svg'
-    }
-  )
-})
+Hooks.once('setup', Setup)
 
 Hooks.on('getHeaderControlsJournalEntrySheet', GetJournalSheetHeaderButtons)
 Hooks.on('getHeaderControlsMacroConfig', GetMacroConfigHeaderButtons)
@@ -124,77 +65,13 @@ Hooks.on('getPlaylistConfigHeaderButtons', GetPlaylistConfigHeaderButtons)
 Hooks.on('getRollTableConfigHeaderButtons', GetRollTableConfigHeaderButtons)
 Hooks.on('getSceneConfigHeaderButtons', GetSceneConfigHeaderButtons)
 
-Hooks.on('createActiveEffect', (data, options, userId) => {
-  if (game.userId === userId) {
-    const statusKey = CoC7ActiveEffect.getStatusKey(data)
-    if (statusKey) {
-      switch (statusKey) {
-        case COC7.status.indefInsane:
-        case COC7.status.unconscious:
-        case COC7.status.criticalWounds:
-        case COC7.status.dying:
-        case COC7.status.prone:
-        case COC7.status.dead:
-          data.parent.setCondition(statusKey, {
-            forceValue: true
-          })
-          break
-        case COC7.status.tempoInsane:
-          {
-            const realTime = data.flags.CoC7?.realTime
-            let duration = null
-            if (realTime === true) {
-              duration = data.duration?.rounds
-            } else if (realTime === false) {
-              duration = data.duration?.seconds
-              if (!isNaN(duration)) {
-                duration = Math.floor(duration / 3600)
-              }
-            }
-            data.parent.setCondition(COC7.status.tempoInsane, {
-              forceValue: true,
-              realTime,
-              duration
-            })
-          }
-          break
-      }
-    }
-  }
-})
+Hooks.on('createActiveEffect', CreateActiveEffect)
 
-Hooks.on('deleteActiveEffect', (data, options, userId) => {
-  if (game.userId === userId) {
-    const statusKey = CoC7ActiveEffect.getStatusKey(data)
-    if (statusKey) {
-      switch (statusKey) {
-        case COC7.status.tempoInsane:
-        case COC7.status.indefInsane:
-        case COC7.status.unconscious:
-        case COC7.status.criticalWounds:
-        case COC7.status.dying:
-        case COC7.status.prone:
-        case COC7.status.dead:
-          data.parent.unsetCondition(statusKey, {
-            forceValue: true
-          })
-      }
-    }
-  }
-})
+Hooks.on('deleteActiveEffect', DeleteActiveEffect)
 
 // This will hide the item called '__CoC7InternalItem__'
 // This item is used for internal purposes and should not be seen by anyone
-Hooks.on('changeSidebarTab', directory => {
-  if (directory instanceof ItemDirectory) {
-    const item = game.items.find(i => i.name === '__CoC7InternalItem__')
-    if (item) {
-      const html = directory._element
-      const itemElement = html.find(`[data-document-id='${item.id}']`)
-      if (itemElement) itemElement[0].style.display = 'none'
-    }
-  }
-})
+Hooks.on('changeSidebarTab', ChangeSidebarTab)
 
 Hooks.on('hotbarDrop', (bar, data, slot) => {
   return CoC7Utilities.createMacro(bar, data, slot)
@@ -207,129 +84,12 @@ Hooks.on('updateChatMessage', (chatMessage, chatData, diff, speaker) =>
   CoC7Chat.onUpdateChatMessage(chatMessage, chatData, diff, speaker)
 )
 
-Hooks.on('ready', async () => {
-  // CONFIG.compatibility.mode = CONST.COMPATIBILITY_MODES.SILENT
-  await Updater.checkForUpdate()
-
-  // game.CoC7.menus = new CoC7Menu();
-
-  activateGlobalListener()
-
-  // setGlobalCssVar()
-
-  game.CoC7.skillList = await game.packs.get('CoC7.skills')?.getDocuments()
-
-  game.socket.on('system.CoC7', async data => {
-    CoC7SystemSocket.callSocket(data)
-  })
-
-  // "SETTINGS.BoutOfMadnessPhobiasIndex": "Phobias index",
-  // "SETTINGS.BoutOfMadnessPhobiasIndexHint": "The index (roll result) that will trigger a roll in the phobias table",
-  // "SETTINGS.BoutOfMadnessManiasIndex": "Manias index",
-  // "SETTINGS.BoutOfMadnessManiasIndexHint": "The index (roll result) that will trigger a roll in the manias table",
-  // "SETTINGS.SamplePhobiasTable": "Sample phobias table",
-  // "SETTINGS.SampleManiasTable": "Sample Manias table",
-
-  function _tableSettingsChanged (table, id) {
-    if (id === 'none') game.CoC7.tables[table] = null
-    else game.CoC7.tables[table] = game.tables.get(id)
-  }
-
-  // function _tableIndexChanged( table, index){
-  //  game.CoC7.tables[table]=index;
-  // }
-
-  const tableChoice = { none: 'SETTINGS.LetKeeperDecide' }
-  for (const t of game.tables) {
-    tableChoice[t._id] = t.name
-  }
-
-  game.settings.register('CoC7', 'boutOfMadnessSummaryTable', {
-    name: 'SETTINGS.BoutOfMadnessSummaryTable',
-    scope: 'world',
-    config: true,
-    default: 'none',
-    type: String,
-    choices: tableChoice,
-    onChange: id => _tableSettingsChanged('boutOfMadness_Summary', id)
-  })
-
-  game.settings.register('CoC7', 'boutOfMadnessRealTimeTable', {
-    name: 'SETTINGS.BoutOfMadnessRealTimeTable',
-    scope: 'world',
-    config: true,
-    default: 'none',
-    type: String,
-    choices: tableChoice,
-    onChange: id => _tableSettingsChanged('boutOfMadness_RealTime', id)
-  })
-
-  // game.settings.register('CoC7', 'boutOfMadnessPhobiasIndex',{
-  //  name: 'SETTINGS.BoutOfMadnessPhobiasIndex',
-  //  hint: 'SETTINGS.BoutOfMadnessPhobiasIndexHint',
-  //  scope: 'world',
-  //  config: true,
-  //  default: 9,
-  //  type: Number,
-  //  onChange:  id => _tableIndexChanged( 'phobiasIndex', id)
-  // });
-
-  // game.settings.register('CoC7', 'boutOfMadnessManiasIndex',{
-  //  name: 'SETTINGS.BoutOfMadnessManiasIndex',
-  //  hint: 'SETTINGS.BoutOfMadnessManiasIndexHint',
-  //  scope: 'world',
-  //  config: true,
-  //  default: 10,
-  //  type: Number,
-  //  onChange:  id => _tableIndexChanged( 'maniasIndex', id)
-  // });
-
-  // game.settings.register('CoC7', 'samplePhobiasTable',{
-  //  name: 'SETTINGS.SamplePhobiasTable',
-  //  scope: 'world',
-  //  config: true,
-  //  default: 'none',
-  //  type: String,
-  //  choices: tableChoice,
-  //  onChange:  id => _tableSettingsChanged( 'phobias', id)
-  // });
-
-  // game.settings.register('CoC7', 'sampleManiasTable',{
-  //  name: 'SETTINGS.SampleManiasTable',
-  //  scope: 'world',
-  //  config: true,
-  //  default: 'none',
-  //  type: String,
-  //  choices: tableChoice,
-  //  onChange:  id => _tableSettingsChanged( 'manias', id)
-  // });
-
-  game.CoC7.tables = {
-    boutOfMadness_Summary:
-      game.settings.get('CoC7', 'boutOfMadnessSummaryTable') === 'none'
-        ? null
-        : game.tables.get(
-          game.settings.get('CoC7', 'boutOfMadnessSummaryTable')
-        ),
-    boutOfMadness_RealTime:
-      game.settings.get('CoC7', 'boutOfMadnessRealTimeTable') === 'none'
-        ? null
-        : game.tables.get(
-          game.settings.get('CoC7', 'boutOfMadnessRealTimeTable')
-        )
-    // maniasIndex: ge.settings.get('CoC7', 'boutOfMadnessPhobiasIndex'),
-    // phobiasIndex: game.settings.get('CoC7', 'boutOfMadnessManiasIndex'),
-    // phobias: ('none' == game.settings.get('CoC7', 'samplePhobiasTable'))?null:game.tables.get(game.settings.get('CoC7', 'samplePhobiasTable')),
-    // manias: ('none' == game.settings.get('CoC7', 'sampleManiasTable'))?null:game.tables.get(game.settings.get('CoC7', 'sampleManiasTable')),
-  }
-})
+Hooks.on('ready', Ready)
 
 // Hooks.on('preCreateActor', (createData) => CoCActor.initToken( createData));
 
-Hooks.on(
-  'renderCoC7ChaseSheet',
-  /** async */ (app, html, data) =>
-    /** await */ CoC7ChaseSheet.setScroll(app, html, data)
+Hooks.on('renderCoC7ChaseSheet', (app, html, data) =>
+  CoC7ChaseSheet.setScroll(app, html, data)
 )
 
 Hooks.on('closeCoC7ChaseSheet', (app, html) =>
@@ -352,28 +112,6 @@ Hooks.on('chatMessage', CoC7Utilities.ParseChatEntry)
 // Hooks.on('createToken', ( scene, actor, options, id) => CoCActor.preCreateToken( scene, actor, options, id))
 // Hooks.on("renderChatLog", (app, html, data) => CoC7Item.chatListeners(html));
 
-Hooks.on('getSceneControlButtons', (/* controls */) => {
-  // if( game.user.isGM){
-  //  let group = controls.find(b => b.name == 'token');
-  //  group.tools.push({
-  //    toggle: true,
-  //    icon : 'fas fa-angle-double-up',
-  //    name: 'devphase',
-  //    active: game.settings.get('CoC7', 'developmentEnabled'),
-  //    title: game.settings.get('CoC7', 'developmentEnabled')? game.i18n.localize( 'CoC7.DevPhaseEnabled'): game.i18n.localize( 'CoC7.DevPhaseDisabled'),
-  //    onClick :async () => await CoC7Utilities.toggleDevPhase()
-  //  });
-  //  group.tools.push({
-  //    toggle: true,
-  //    icon : 'fas fas fa-user-edit',
-  //    name: 'charcreate',
-  //    active: game.settings.get('CoC7', 'charCreationEnabled'),
-  //    title: game.settings.get('CoC7', 'charCreationEnabled')? game.i18n.localize( 'CoC7.CharCreationEnabled'): game.i18n.localize( 'CoC7.CharCreationDisabled'),
-  //    onClick :async () => await CoC7Utilities.toggleCharCreation()
-  //  });
-  // }
-})
-
 // Hooks.on('renderSceneControls', () => CoC7Utilities.updateCharSheets());
 // Hooks.on('renderSceneNavigation', () => CoC7Utilities.updateCharSheets());
 // Sheet css options
@@ -387,17 +125,6 @@ Hooks.on('renderSceneControls', CoC7Menu.renderControls)
 Hooks.on('dropCanvasData', CoC7Canvas.onDropSomething)
 
 Hooks.on('dropActorSheetData', DropActorSheetData)
-
-function activateGlobalListener () {
-  const body = $('body')
-  document.addEventListener('click', CoC7ContextMenu.closeAll)
-  body.on('click', 'a.coc7-inline-check', CoC7Check._onClickInlineRoll)
-  document.addEventListener('mousedown', _onLeftClick)
-}
-
-function _onLeftClick (event) {
-  return event.shiftKey
-}
 
 Hooks.on('drawNote', DrawNote)
 Hooks.on('renderActorDirectory', RenderActorDirectory)
