+++ templates/common/active-effects.hbs
@@ -27,7 +27,7 @@
           <li class="item effect flexrow-coc7" data-effect-id="{{effect.id}}">
             <div class="item-name effect-name flexrow-coc7">
               <img class="item-image" src="{{effect.icon}}"/>
-              <h4>{{#if effect.name}}{{effect.name}}{{!-- // FoundryVTT v10 --}}{{else}}{{effect.label}}{{/if}}</h4>
+              <h4>{{effect.name}}</h4>
             </div>
             <div class="effect-source">{{effect.sourceName}}</div>
             <div class="effect-duration">{{effect.duration.label}}</div>
